+++ src/gui/state/game/detection/heroic.cpp
@@ -99,7 +99,7 @@ std::filesystem::path GetUserConfigPath() {
   PWSTR path;
 
   if (SHGetKnownFolderPath(FOLDERID_RoamingAppData, 0, NULL, &path) != S_OK) {
-    throw std::system_error(GetLastError(),
+    throw std::system_error(static_cast<int>(GetLastError()),
                             std::system_category(),
                             "Failed to get %APPDATA% path.");
   }
